Move log and stdout to inner folder
